Added about us page and updated design
I fixed the split paragrah issue.

Highlighted the 'did you knows'.

Fixed the Next button to appear when an article has no questions at the bottom.

Also the Next button no longer appears if it reaches the last available article in the database.

Added an about us page (which only has the bit from the home page so far).

Worked a bit on the layout in general.

Created basic layout for map page

Updated maps, graphs and text